+++ app/Http/Controllers/Admin/EventMaintenance/FundSourceMaintenanceController.php
@@ -13,36 +13,41 @@ use App\Services\Admin\EventMaintenance\FundSource\{
     FundSourceUpdateService,
     FundSourceDeleteService,
 };
+use App\Services\PermissionServices\PermissionCheckingService;
 
 use App\Http\Controllers\Controller as Controller;
 
 class FundSourceMaintenanceController extends Controller
 {
     protected $viewDirectory = 'admin.maintenances.event.fundSource.';
+    protected $permissionChecker;
 
     /**
      * Create a new controller instance.
-     *
      * @return void
      */
     public function __construct()
     {
         $this->middleware('auth');
+        $this->permissionChecker = new PermissionCheckingService();
     }
 
     public function index()
     {
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-Super-Admin-Manage_Event'), 403);
         $fundSources = FundSource::all();
         return view($this->viewDirectory . 'index', compact('fundSources',));
     }
     
     public function create()
     {
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-Super-Admin-Manage_Event'), 403);
         return view($this->viewDirectory . 'create',);
     }
 
     public function store(FundSourceStoreRequest $request)
     {
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-Super-Admin-Manage_Event'), 403);
         $message = (new FundSourceStoreService())->store($request);
 
         return redirect()->action(
@@ -52,6 +57,7 @@ class FundSourceMaintenanceController extends Controller
 
     public function edit($fund_source_id)
     {
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-Super-Admin-Manage_Event'), 403);
         $fundSource = $this->checkIfFundSourceExists($fund_source_id);
 
         return view($this->viewDirectory . 'edit', compact('fundSource'));
@@ -59,6 +65,7 @@ class FundSourceMaintenanceController extends Controller
 
     public function update(FundSourceUpdateRequest $request, $fund_source_id)
     {
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-Super-Admin-Manage_Event'), 403);
         $fundSource = $this->checkIfFundSourceExists($fund_source_id);
 
         $message = (new FundSourceUpdateService())->update($fundSource, $request);
@@ -71,6 +78,7 @@ class FundSourceMaintenanceController extends Controller
 
     public function show($fund_source_id)
     {
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-Super-Admin-Manage_Event'), 403);
         $fundSource = $this->checkIfFundSourceExists($fund_source_id);
         
         return view($this->viewDirectory . 'show', compact('fundSource'));
@@ -78,6 +86,7 @@ class FundSourceMaintenanceController extends Controller
 
     public function destroy(FundSourceDeleteRequest $request, $fund_source_id)
     {
+        abort_if(! $this->permissionChecker->checkIfPermissionAllows('AR-Super-Admin-Manage_Event'), 403);
         $fundSource = $this->checkIfFundSourceExists($fund_source_id);
 
         $message = (new FundSourceDeleteService())->delete($fundSource, $request);
